Update ParameterBindingData to follow CloudEvent pattern
