+++ src/Microsoft.Azure.WebJobs/WebJobs.csproj
@@ -25,6 +25,7 @@
     </PackageReference>
     <PackageReference Include="System.ComponentModel.Annotations" Version="4.4.0" />
     <PackageReference Include="System.Diagnostics.TraceSource" Version="4.3.0" />
+    <PackageReference Include="System.Memory.Data" Version="1.0.1" />
   </ItemGroup>
 
   <ItemGroup>
